+++ app/Models/User.php
@@ -17,6 +17,7 @@ use Illuminate\Notifications\Notifiable;
  * @property \Illuminate\Support\Carbon|null $email_verified_at
  * @property string $password
  * @property string|null $remember_token
+ * @property string $timezone
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
  * @property-read \Illuminate\Notifications\DatabaseNotificationCollection|\Illuminate\Notifications\DatabaseNotification[] $notifications
@@ -58,6 +59,7 @@ class User extends Authenticatable implements FilamentUser
 		'name',
 		'email',
 		'password',
+		'timezone',
 	];
 
 	/**